Moved validator class to a single method in messagehandler
